v1.15
new: animated quicksaves
new: tv grain option (ZX_GRAIN)
new: visual effects while holding reset button
new: can remap individual gamepad buttons now
new: added italian>english game menu translations (@koseid72)
new: releases contain AMR64 binaries now (macos)
new: results in search tab use standard list views now
chg: darker default palette
chg: disabled gamepads on linux for now (buggy on virtualbox+lubuntu16 at least)
chg: do not wipe medias if ZX_AUTOLOAD is off (go fully manual instead)
chg: relocated fullscreen close button to top-right corner
fix: case where filtering on local files would also search mistakenly in paths
fix: do not send enter vk to zx while pressing alt+enter (3rd_tigr) (linux,macos)
fix: crash while transcoding very large text sequences (>10kib) (see: mythhistory instructions)
fix: missed to serialize ZX_TAPECOUNTER option
fix: recreated fullscreen window during launch (3rd_tigr) (linux)
fix: removed local "games/" folder scan in old game browser (when ZXDB is disabled)
fix: removed zxdb watermarks from linux and macos binaries
fix: slightly incorrect alphas and/or colors when filtering zxdb titles
fix: clicking on tape browser, which was broken since flashload was introduced

+++ src/zx_tap.h
@@ -377,7 +377,7 @@ int tape_seeki(int at) {
     if(ZX_FLASHLOAD) if(TAP_sof) return 0;
     return voc_pos = voc && voc_len && at >= 0 && at < voc_len ? at : voc_pos;
 }
-int tape_seekf(int at) {
+int tape_seekf(float at) {
     if(ZX_FLASHLOAD) if(TAP_sof) return 0;
     return voc_pos = voc && voc_len && at >= 0 && at <= 1 ? at * (voc_len - 1) : voc_pos;
 }
